Provide path and line to :Gbrowse providers for commits
Look for and use path of the +++ side when it exists. For removals it is
/dev/null, so use path of the  --- side. For the line range only line1
is set, corresponding to the first line of selected region. If this line
is in the removed part of the hunk, the number is relative to the ---
side and is negative to indicate that to adapters.