Extend :Gbrowse to support line anchors for commits
In commit buffers :Gbrowse with range will include 'start' and
'end' dictionaries to metadata passed to upstram providers containing
fields:

* 'cursor': ' ', '-' or '+' when position is on the context, removed or
  added line
* 'oldpath': path of the '---' side of the diff hunk (empty for
  additions)
* 'newpath': path of the '+++' side of the diff hunk (empty for
  removals)
* 'oldline': line number for the '---' side
* 'newline': line number for the '+++' side

For positions in added or removed lines, opposite side line number is
taken as for the *next* context line (even if it doesn't exist).
Exception is completely missing paths (additions or removals) -
corresponding line number is always 0. This is the same numbering logic
as in the diff hunk header.